requirements.txt
requests
beautifulsoup4
notion-client
python-dotenv 